[Crear endpoint para administrador]
- Eliminado el comando CLI para la creación de usuarios administradores.
- Añadido un endpoint POST /api/create-admin en routes.py para permitir la creación de un administrador de prueba directamente desde el backend.

Este cambio facilita la configuración inicial en entornos temporales o de desarrollo.
