+++ src/api/commands.py
@@ -32,43 +32,3 @@ def setup_commands(app):
     @app.cli.command("insert-test-data")
     def insert_test_data():
         pass
-
-    
-    # -----------------------------------------------------------------
-    # ADMIN USER CREATION COMMAND
-    # -----------------------------------------------------------------
-    @app.cli.command("create-admin")
-    def create_admin():
-        """
-        Comando para crear un usuario administrador.
-        Este comando permite la creación de un usuario administrador predeterminado en el sistema.
-        """
-        from flask_bcrypt import Bcrypt
-        bcrypt = Bcrypt()
-
-        print("Starting the creation of an administrator user...")
-
-        admin_email = "admin@anda.com.uy"
-        admin_password = "adminpass"
-
-        with app.app_context():
-            existing_admin = User.query.filter_by(email=admin_email).first()
-            if existing_admin:
-                print(f"Admin user with email {admin_email} already exists. Skipping creation.")
-                return
-
-            admin = User(
-                user_name="admin",
-                email=admin_email,
-                password_hash=bcrypt.generate_password_hash(admin_password).decode('utf-8'),
-                role="admin",
-                status="activo"
-            )
-
-            db.session.add(admin)
-            try:
-                db.session.commit()
-                print(f"Administrator user created successfully with email: {admin_email}")
-            except Exception as e:
-                db.session.rollback()
-                print(f"Error occurred while creating the administrator user: {e}")
